+++ app.js
@@ -826,6 +826,7 @@ import { renderers as RV1470 } from './views/group-v1470.js';
 import { renderers as RV1471 } from './views/group-v1471.js';
 import { renderers as RV1472 } from './views/group-v1472.js';
 import { renderers as RV1473 } from './views/group-v1473.js';
+import { renderers as RV1475 } from './views/group-v1475.js';
 import { renderers as RV1396 } from './views/group-v1396.js';
 import { renderers as RV164 } from './views/group-v164.js';
 import { renderers as RV165 } from './views/group-v165.js';
@@ -914,7 +915,7 @@ const RENDERERS = { ...RA, ...RB, ...RC, ...RE, ...RF, ...RG, ...RH, ...RI, ...R
   ...RV902,
   ...RV903,
   ...RV905,
-  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV63, ...RPALINT };
+  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV1475, ...RV63, ...RPALINT };
 
 // ----- Utility registry ----------------------------------------------------
 // Source of truth for routes, names, group, audiences, and clinical flag.
@@ -1909,6 +1910,7 @@ const UTILITIES = [
   { id: 'poseidon',                 name: 'POSEIDON Classification (Low-Prognosis Patients in ART)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'glass-stage',              name: 'GLASS Anatomic Stage (Chronic Limb-Threatening Ischemia)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'irecist',                  name: 'iRECIST Time-Point Response (Immunotherapy Trials)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
+  { id: 'mrecist', name: 'mRECIST Response (Hepatocellular Carcinoma)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'snot22',                   name: 'SNOT-22 (Sino-Nasal Outcome Test)', group: 'G', audiences: ['clinicians', 'patients', 'educators'], clinical: true },
   { id: 'puqe24',                   name: 'PUQE-24 (Nausea and Vomiting of Pregnancy)', group: 'G', audiences: ['clinicians', 'patients', 'educators'], clinical: true },
   { id: 'gags',                     name: 'GAGS (Global Acne Grading System)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },